backend/requirements.txt
Flask
Flask-SQLAlchemy
pandas
Flask-Cors
python-dotenv
click